+++ components/layout/Header.module.css
@@ -8,6 +8,7 @@
 	align-items: center;
 	justify-content: space-evenly;
 	padding: 1rem;
+	flex-wrap: wrap;
 
 }
