+++ components/layout/navigation/NavItem.js
@@ -1,0 +1,11 @@
+import MenuItem from "@mui/material/MenuItem";
+import ListItemText from "@mui/material/ListItemText";
+import ListItemIcon from "@mui/material/ListItemIcon";
+export default function NavItem({ icon, text, onClick }) {
+  return (
+    <MenuItem onClick={onClick}>
+      <ListItemIcon></ListItemIcon>
+      <ListItemText></ListItemText>
+    </MenuItem>
+  );
+}
